refactor: Setting内カテゴリをTemplateと同様のリンク配置に変更 (#34)
- SettingsPageのCategorySectionインライン表示をActionsセクション内のリンクボタンに変更
- CategoriesPage・/categoryルートを復活
- 不要になったCategorySection.tsxを削除
