+++ frontend/src/router.tsx
@@ -2,6 +2,7 @@ import { Navigate, createBrowserRouter } from "react-router"
 
 import { AppLayout } from "./components/AppLayout"
 import { ProtectedRoute } from "./components/ProtectedRoute"
+import { CategoriesPage } from "./pages/CategoriesPage"
 import { ExpenseAnnualPage } from "./pages/ExpenseAnnualPage"
 import { ExpenseDetailPage } from "./pages/ExpenseDetailPage"
 import { ExpenseMonthlyPage } from "./pages/ExpenseMonthlyPage"
@@ -35,6 +36,7 @@ export const router = createBrowserRouter([
           { path: ":uuid", element: <ExpenseDetailPage /> },
         ],
       },
+      { path: "/category", element: <CategoriesPage /> },
       { path: "/setting", element: <SettingsPage /> },
     ],
   },
